requirements.txt
fastapi>=0.110.0
pydantic>=2.7.0
pydantic_core>=2.18.0
SQLAlchemy>=2.0.25
uvicorn>=0.29.0
python-multipart>=0.0.9
scikit-learn>=1.3.0
pandas>=2.0
google-genai>=1.21.1
ipykernel>=6.29.5
nbformat>=5.10.4
openai>=1.90.0
plotly>=6.1.2
python-dotenv>=1.1.0
requests>=2.32.4
chromadb>=0.5.0
numpy<2.0
sentence-transformers>=2.2.2
kagglehub>=0.2.6
spacy>=3.8.8
prometheus-client>=0.23.1
python-json-logger>=4.0.0
pydantic[email]
langchain-google-genai
Pillow>=10.0.0
langchain>=0.2.11
langchain-core>=0.2.38
langdetect>=1.0.9